+++ examples/rust_client/target/debug/tinybit-example-client.d
@@ -1,0 +1,1 @@
+/home/user/tinybit/examples/rust_client/target/debug/tinybit-example-client: /home/user/tinybit/examples/rust_client/src/main.rs